# Backend/java/src/main/resources/static/swagger.yaml
openapi: "3.0.3"
info:
  title: "NoteS API"
  version: "v 0.0.1"
servers:
  - url: "{scheme}://{domain}:{port}/{context}"
    description: Resource server
    variables:
      scheme:
        enum:
          - 'http'
          - 'https'
        default: 'http'
      domain:
        enum:
          - 'localhost'
          - 'notes.fvds.ru'
        default: 'localhost'
      port:
        enum:
          - '8080'
          - '9080'
          - '9180'
          - '5267'
          - '8443'
          - '80'
          - '443'
        default: '5267'
      context:
        enum:
          - ''
          - 'server'
        default: ''
components:
  schemas:
...
    CommentCreateRequestDto:
      required:
        - syntax_name
      type: object
      properties:
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        syntax_name:
          "$ref": "#/components/schemas/SyntaxTypeName"
        content:
          type: string
          nullable: true
      additionalProperties: false
    CommentCreateResponseDto:
      required:
        - note_type
        - syntax_name
      type: object
      properties:
        path:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        note_type:
          "$ref": "#/components/schemas/NoteTypeName"
        is_public:
          type: boolean
        syntax_name:
          "$ref": "#/components/schemas/SyntaxTypeName"
        content:
          type: string
          nullable: true
      additionalProperties: false
    CommentEditRequestDto:
      required:
        - syntax_name
      type: object
      properties:
        title:
          type: string
          nullable: true
        syntax_name:
          "$ref": "#/components/schemas/SyntaxTypeName"
        content:
          type: string
          nullable: true
      additionalProperties: false
    CommentEditResponseDto:
      required:
        - note_type
        - syntax_name
      type: object
      properties:
        path:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        note_type:
          "$ref": "#/components/schemas/NoteTypeName"
        is_public:
          type: boolean
        syntax_name:
          "$ref": "#/components/schemas/SyntaxTypeName"
        content:
          type: string
          nullable: true
      additionalProperties: false
...
    NoteEditRequestDto:
      type: object
      properties:
        content:
          type: string
          nullable: true
        syntax_name:
          "$ref": "#/components/schemas/SyntaxTypeName"
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
      additionalProperties: false
...
    NoteSearchContentResponseDto:
      required:
        - note_type
      type: object
      properties:
        title:
          type: string
          nullable: true
        path:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        owner_account_name:
          type: string
          nullable: true
        main_note_path:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        score:
          type: number
          format: double
        note_type:
          "$ref": "#/components/schemas/NoteTypeName"
        tags:
          type: array
          items:
            "$ref": "#/components/schemas/TagResponseDto"
          nullable: true
        is_public:
          type: boolean
        created_at:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
    NoteSearchContentResponseDtoPageDto:
      required:
        - items
        - total
        - total_pages
      type: object
      properties:
        items:
          type: array
          items:
            "$ref": "#/components/schemas/NoteSearchContentResponseDto"
          nullable: true
        total_pages:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int64
      additionalProperties: false
    NoteTypeName:
      enum:
        - NOTE
        - COMMENT
        - COMMENT_REDACTED
      type: string
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
    SyntaxTypeName:
      enum:
        - PLAINTEXT
        - MARKDOWN
      type: string
    TagResponseDto:
      type: object
      properties:
        name:
          type: string
          nullable: true
        color:
          maximum: 16777215
          minimum: 0
          type: integer
          format: int32
      additionalProperties: false
  securitySchemes:
    Bearer:
      type: oauth2
      description: API использует [OAuth2 authorizationCode](https://auth0.com/docs/get-started/authentication-and-authorization-flow/authorization-code-flow).
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8080/realms/NoteS/protocol/openid-connect/auth
          tokenUrl: http://localhost:8080/realms/NoteS/protocol/openid-connect/token
          scopes:
            read: Balea Server HTTP Api
security:
  - Bearer: [ ]
paths:
...
  "/api/public/{accountName}/notes/{pathNote}/comments":
    get:
      tags:
        - PublicComments
      description: Список комментариев к заметке
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: pathNote
          in: path
          required: true
          schema:
            maxLength: 128
            minLength: 3
            type: string
        - name: Page
          in: query
          schema:
            maximum: 2147483647
            minimum: 0
            type: integer
            format: int32
        - name: Limit
          in: query
          schema:
            maximum: 20
            minimum: 5
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/NoteSearchContentResponseDtoPageDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/NoteSearchContentResponseDtoPageDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/NoteSearchContentResponseDtoPageDto"
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
    post:
      tags:
        - PublicComments
      description: Создание комментария к заметке
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: pathNote
          in: path
          required: true
          schema:
            maxLength: 128
            minLength: 3
            type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CommentCreateRequestDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/CommentCreateRequestDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/CommentCreateRequestDto"
      responses:
        '201':
          description: Created
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/CommentCreateResponseDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/CommentCreateResponseDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/CommentCreateResponseDto"
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
  "/api/public/{accountName}/comments/{pathNote}":
    post:
      tags:
        - PublicComments
      description: Редактирование комментария
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: pathNote
          in: path
          required: true
          schema:
            maxLength: 128
            minLength: 3
            type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CommentEditRequestDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/CommentEditRequestDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/CommentEditRequestDto"
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/CommentEditResponseDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/CommentEditResponseDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/CommentEditResponseDto"
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
    delete:
      tags:
        - PublicComments
      description: Удаление комментария
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: pathNote
          in: path
          required: true
          schema:
            maxLength: 128
            minLength: 3
            type: string
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
...
  "/api/public/{accountName}/notes/{pathNote}":
    post:
      tags:
        - PublicNote
      description: Редактирование заметки, комментарий в другом месте.
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: pathNote
          in: path
          required: true
          schema:
            maxLength: 128
            minLength: 3
            type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/NoteEditRequestDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/NoteEditRequestDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/NoteEditRequestDto"
      responses:
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
    delete:
      tags:
        - PublicNote
      description: Удаление заметки
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: pathNote
          in: path
          required: true
          schema:
            maxLength: 128
            minLength: 3
            type: string
      responses:
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
    get:
      tags:
        - PublicNote
      description: Данные заметки или комментария
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: pathNote
          in: path
          required: true
          schema:
            maxLength: 128
            minLength: 3
            type: string
      responses:
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
...